Created the relax_io.readlines method for handling the bz2 and gzip module byte strings in Python 3.
The bz2.BZ2File and gzip.GzipFile classes now return lists of bytes strings from their readlines()
methods in Python 3.  This causes failures all over relax.  This behaviour is probably a bug in
these classes.  The relax_io.readlines method is a workaround for this problem, to have a list of
standard strings returned.