Redesigned the reset user function backend.
This now no only clears out the relax data store, but it also resets the GUI if present.  Some of
the reset code comes from the tearDown() method of the GUI tests.  All windows but the main GUI
window are closed and the relax controller gauges are set to zero and the log window text cleared.

These changes should allow GUI tests after an error or failure to pass, something which is currently
problematic.