Big fixes for the running of the GUI tests.
The GUI initialisation and termination, when not running in GUI mode, is now preformed outside of
the Python unittest framework.  The GUI was previously initalised in the setUp() method for all GUI
tests, but for some reason this caused abrupt terminations of the GUI_test_runner.run() method.
This run method, when finished, would not return to the Test_suite_runner but would cause the Python
interpreter to silently die.  The result was that the test suite summary would appear not to be
printed out.