Fix for bug #20480 (https://gna.org/bugs/?20480).
This was reported by Stanislava Panova (https://gna.org/users/stacy).

The problem was that the state file was saved after a RelaxError occurred during the model-free
analysis.  The GUI data store information states that the active data pipe is called 'final'.  But
an error occurred prior to the 'final' data pipe being created.  The fix was to allow the state to
be loaded even if the GUI data store information has been corrupted.