The structure.find_pivot user function now uses the logarithmic barrier function.
This is for constrained optimisation and requires the newest minfx code.  The pivot position is
constrained within a box of +/- 1000 Angstroms from zero.  This is needed for when the solution is
an infinite line - i.e. a rotation axis and not a pivot point.  Previously the simplex optimisation
would head toward + or - infinity.  But now with a logarithmic barrier, the simplex algorithm can
stabilise and find a point on the axis very quickly, long before reaching the edges of the box.