Large cleanup and bugfixes for the molecule, residue, and spin data structure metadata maintenance.
The bugs fixed are important for non-protein molecules.  For example is the spin name is not unique
per residue, or per molecule if no residues are defined, many parts of relax would fail.

All of the metadata_*() and spin_id_variants*() functions have been redesigned.  It was also
identified that metadata_prune() was being used by different parts of relax for two different
purposes - the removal or pruning of metadata prior to the deletion of a data structure and the
clean up of no longer valid metadata.  These two goals conflicted resulting in unpredictable
behaviour.  Therefore the new metadata_cleanup() and spin_id_variants_cleanup() functions have been
created and the two behaviours separated.