Fix for a nasty bug introduced at r18901.
The string decode() method was being called for the version.url() and version.revision() functions
to handle Python 3.  However this was causing strings in Python 2 to be converted to unicode, which
then caused the saving of XML state files to fail.  Now the version information is explicitly
converted to a string using the str() builtin function.