Shifted the standard deviation code from the Monte Carlo error_analysis() function to the lib package.
The new function lib.stats.std() is now used to simplify the error_analysis() function and allow the
code to be reused.  This will be useful for expanding the pipe_control.monte_carlo.error_analysis()
function to handle parameters which are dictionaries, for example as in the relax_disp branch.