Shifted and merged the NOE and Rx peak intensity wizards into a new module.
The wizards were separate and a part of the analysis frame class objects.  The two wizards have been
merged into the gui.wizards.peak_intensity.Peak_intensity_wizard class as most of the code is
shared.  This one wizard class will be useful for reusing in the relaxation dispersion branch.