Fix for the nasty bug #20756 (https://gna.org/bugs/?20756).
The problem was that the global execution lock was not always released by a relax script when
certain errors are raised during the script execution.  This does not occur for all types of error
though.  Now the release of the lock has been shifted into the 'finally' statement to absolutely
force lock release.