Fix for the package checking as part of the unit tests.
This was identified from the bug report #20820 (https://gna.org/bugs/?20820) submitted by Troels E.
Linnet (https://gna.org/users/tlinnet).

The problem was that on some systems, the full path is required for checking the presence of the
directories which are the sub-packages of the main package being checked.  The result was that 
checking for the package in the __all__ list was skipped.

Note that this change does not fix the bug reported.