Fix for bug #20918 (https://gna.org/bugs/?20918), the hanging of the data pipe editor.
This was reported by Troels Linnet (https://gna.org/users/tlinnet) and is an MS Windows only
problem.

The problem is in the wxMSW part of wxPython, and it may be fixed in newer wxPython versions.  The
issue is nevertheless now avoided by calling the GUI user function store objects with the arguments
wx_wizard_sync=True and wx_wizard_modal=True.  This appears to solve the problem.