Modified the Monte Carlo simulation printout behaviour for the minimisation related user functions.
This is to help in fixing bug #21190 (https://gna.org/bugs/?21190).

This includes the calculate, grid_search, and minimise user functions.  The new multi-processor
is_queued() method is used to determine if the optimisation code of the specific analysis has queued
rather than run the calculations.  If queued, the 'Simulation X' text will not be printed out.  This
avoids printing out all the text at the start before anything has happened.  The specific
multi-processor optimisation code must provide it's own printouts when each calculation is complete.