Expansion and fixes for the Relax_disp.test_hansen_cpmg_data_missing_auto_analysis system test.
The parameters for spin :4 are now being checked, and all the checks updated for the changed data.
The parameter values are slightly different as data is now missing and because only 3 spins are used
for the error analysis whereas in all other Hansen CPMG data sets the more accurate errors are from
all spins.