Another fix for bug #21309, the 'R2eff' dispersion model failure when peak intensity data is missing.
The bug is reported at https://gna.org/bugs/?21309.

This second problem is only for the numeric CPMG models for when all data at one magnetic field
strength is missing.  When the relaxation dispersion target function is being set up, the creation
of the self.power data structure holding the number of CPMG blocks fails.  The problem is that the
relaxation time for the missing field strength is set to NaN.  This is now caught using
lib.float.isNaN().