Removed the 'CR72' model for GUARDD from the dispersion software comparison table.
This is for the dispersion chapter of the user manual.  This software, like cpmg_fit, only supports
the 'MMQ CR72' model which gives slightly different results to the original 'CR72' model when using
only SQ CPMG-type data.  Hence supporting 'MMQ CR72' does not automatically mean that the 'CR72'
model can be optimised.