Better support for the writing out of multiple molecules by the structure.write_pdb user function.
This is for the internal structural object write_pdb() method.  Now each molecule is assigned a
different chain ID in the PDB file, and the chain IDs loaded into the structural object are ignored.
The chain IDs should however be preserved when using structure.read_pdb followed by
structure.write_pdb, without storing the ID.

A number of the Structure system tests had to be updated, as now the relax generated PDB files will
always write out a chain ID.